Extract compareVersions to a pure module; make DB-only boundary transitively true
Addresses the transitive-import boundary hole (external GPT-5.6 review, finding
#3): export.ts imported compareVersions from query.ts, which imports db.ts and
opens the global database at import time — so importing export.ts under an invalid
DB_PATH threw, and the "reads only the passed handle" claim was only true at
runtime, not for the module graph.

- New src/version-compare.ts holds compareVersions as a dependency-free module
  (no bun:sqlite / db.ts / paths.ts). query.ts imports and re-exports it, so
  `from "./query.ts"` callers are unchanged; export.ts imports it directly.
- Boundary test strengthened: the import allowlist now points at
  ./version-compare.ts and additionally asserts that module imports nothing
  (closing the transitive closure), plus a behavioral test that importing
  export.ts under a nonexistent DB_PATH opens no database.

Documents the known compareVersions prerelease-number limitation (7.24beta1 vs
7.24beta2 compare equal) inline; tracked for a shared-comparator fix in the
E1/E2 follow-up issue.
